Added classes for validation of config file
Added ConfigElement class for tree validation of the configuration file. It allows to store all configuration messages (warnings/errors) for each property and section of the configuration file in a tree data structure and implements logic to loop this data structure to build a composite validation message with all validation messages in one overview.

Added ValidationMessage class to store a validation message (warning/error).

Added ValidationLevel enum to distinguish between types of ValidationMessage's. Currently WARNING and ERROR are supported.